fix: fix broken tests
