+++ pkg/controllers/namespaced/verifier_controller_test.go
@@ -24,13 +24,12 @@ import (
 	configv1beta1 "github.com/deislabs/ratify/api/v1beta1"
 	"github.com/deislabs/ratify/internal/constants"
 	"github.com/deislabs/ratify/pkg/controllers"
-	"github.com/deislabs/ratify/pkg/controllers/test"
 	"github.com/deislabs/ratify/pkg/customresources/verifiers"
+	test "github.com/deislabs/ratify/pkg/utils"
 	"github.com/sirupsen/logrus"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
-	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
@@ -295,29 +294,6 @@ func TestVerifierReconcile(t *testing.T) {
 				t.Fatalf("Verifier map expected size %v, actual %v", tt.expectedVerifierCount, controllers.StoreMap.GetStoreCount())
 			}
 		})
-	}
-}
-
-func TestVerifierSetupWithManager(t *testing.T) {
-	scheme, err := test.CreateScheme()
-	if err != nil {
-		t.Fatalf("CreateScheme() expected no error, actual %v", err)
-	}
-	client := fake.NewClientBuilder().WithScheme(scheme)
-	r := &VerifierReconciler{
-		Scheme: scheme,
-		Client: client.Build(),
-	}
-
-	mgr, err := ctrl.NewManager(ctrl.GetConfigOrDie(), ctrl.Options{
-		Scheme: scheme,
-	})
-	if err != nil {
-		t.Fatalf("NewManager() expected no error, actual %v", err)
-	}
-
-	if err := r.SetupWithManager(mgr); err != nil {
-		t.Fatalf("SetupWithManager() expected no error, actual %v", err)
 	}
 }
